requirements.txt
# PyTorch - Install based on your system
# For CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For CPU only: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0

# OpenCLIP - Install from GitHub main branch for MetaCLIP2 support
# pip install --upgrade "open-clip-torch @ git+https://github.com/mlfoundations/open_clip.git@main"
open-clip-torch>=2.20.0

# FAISS - Choose based on your system
# For GPU support: faiss-gpu>=1.7.4
# For CPU only: faiss-cpu>=1.7.4
faiss-cpu>=1.7.4

# Core ML packages
numpy>=1.23.0

# Data processing
pandas>=1.5.0
pyarrow>=10.0.0
pillow>=9.0.0

# Utilities
requests>=2.28.0
tqdm>=4.64.0

